Initial Commit
Basic environment has been setup, I am using HTML, CSS, and
JavaScript. I managed to setup the display of total marks, and some
JavaScript which recognises if the addition arithmetic is correct if the
user presses Enter on the keyboard, which then will increase the marks
by the question value.

I need to work on randomising the first value and the second value after
a question has been answered correctly. Later I can work on the
randomising the arithmetic type to either 'Addition', 'Subtraction',
Multiplication', or 'Division'. This needs to be scaled by difficulty as
the player progresses.