Implementation of Fundamentals
Now both of the values will be randomised after a correct answer,
followed by a sound effect. The display of the marks have been altered,
an extra variable has been introduced 'marksDisplayed'. This variable
will Linear Interpolate to equal the 'marks'. A satisfying effect which
will work even if the distance between the two numbers is large. This is
done in a game loop using the 'window.requestAnimationFrame()' function,
sourced from:

https://spicyyoghurt.com/tutorials/html5-javascript-game-development/create-a-proper-game-loop-with-requestanimationframe