💄 Added images to upgrades, and colour change on hover
The different images will be shown on the left side of the upgrade based on the name of the upgrade.

The upgrade will also change colour to a slightly brighter one when the mouse is hovering over it. This was done using the &:hover pseudo-class with Sass.

Next I need to add functionality for clicking on the upgrade. And also it should flash when purchased.